add code to insert category into database with tests

+++ ime-server/src/main/java/com/scurtis/ime/repository/CategoryRepository.java
@@ -3,5 +3,5 @@ package com.scurtis.ime.repository;
 import com.scurtis.ime.entity.Category;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
-public interface CategoryRepository extends ReactiveCrudRepository<Category, Integer> {
+public interface CategoryRepository extends ReactiveCrudRepository<Category, Long> {
 }
